+++ stocks/src/main/java/com/pyamsoft/tickertape/stocks/network/NetworkStock.kt
@@ -23,10 +23,10 @@ internal data class NetworkStock
 internal constructor(
     internal val symbol: String,
     internal val shortName: String?,
-    internal val regularMarketPrice: Float?,
-    internal val regularMarketChange: Float?,
-    internal val regularMarketChangePercent: Float?,
-    internal val postMarketPrice: Float?,
-    internal val postMarketChange: Float?,
-    internal val postMarketChangePercent: Float?,
+    internal val regularMarketPrice: Double?,
+    internal val regularMarketChange: Double?,
+    internal val regularMarketChangePercent: Double?,
+    internal val postMarketPrice: Double?,
+    internal val postMarketChange: Double?,
+    internal val postMarketChangePercent: Double?,
 )
